+++ src/bit_vectors/darray/mod.rs
@@ -1,5 +1,0 @@
-//! Index types implementing the dense array technique.
-
-pub mod inner;
-
-pub use inner::DArrayIndex;